+++ src/main/java/nl/clockwork/ebms/dao/mssql/EbMSDAOImpl.java
@@ -36,7 +36,7 @@ public class EbMSDAOImpl extends nl.clockwork.ebms.dao.mysql.EbMSDAOImpl
 	}
 
 	@Override
-	public String getMessageIdsQuery(String messageContextFilter, EbMSMessageStatus status, long maxNr)
+	public String getMessageIdsQuery(String messageContextFilter, EbMSMessageStatus status, int maxNr)
 	{
 		return "select top " + maxNr + " message_id" +
 		" from ebms_message" +
