+++ src/main/java/nl/clockwork/ebms/dao/oracle/EbMSDAOImpl.java
@@ -33,7 +33,7 @@ public class EbMSDAOImpl extends AbstractEbMSDAO
 	}
 
 	@Override
-	public String getMessageIdsQuery(String messageContextFilter, EbMSMessageStatus status, long maxNr)
+	public String getMessageIdsQuery(String messageContextFilter, EbMSMessageStatus status, int maxNr)
 	{
 		return "select * from (" +
 		"select message_id" +
