+++ src/main/java/nl/clockwork/ebms/dao/spring/mssql/EbMSDAOImpl.java
@@ -56,7 +56,7 @@ public class EbMSDAOImpl extends nl.clockwork.ebms.dao.spring.mysql.EbMSDAOImpl
 	}
 
 	@Override
-	public String getMessageIdsQuery(String messageContextFilter, EbMSMessageStatus status, long maxNr)
+	public String getMessageIdsQuery(String messageContextFilter, EbMSMessageStatus status, int maxNr)
 	{
 		return "select top " + maxNr + " message_id" +
 		" from ebms_message" +
